Do not allow matmul serialisation factor to be <= 0 when doing serialsation.
Summary: As title.

Test Plan: CI

Reviewers: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Reviewed By: #popart, #framework_ip_review_-_any_oss_or_third-party_code_use_has_been_approved, fabiant

Maniphest Tasks: T45645

Differential Revision: https://phabricator.sourcevertex.net/D51113

+++ willow/include/popart/op/matmul.hpp
@@ -25,7 +25,7 @@ public:
     enum class Mode { None, InputChannels, ReducingDim, OutputChannels };
 
     Mode mode           = Mode::None;
-    int64_t factor      = 0;
+    uint32_t factor     = 0;
     bool keep_precision = false;
   };
